requirements.txt
asgiref==3.4.1
astroid==2.6.2
Django==3.2.5
django-environ==0.4.5
djangorestframework==3.12.4
gunicorn==20.1.0
isort==5.9.1
lazy-object-proxy==1.6.0
mccabe==0.6.1
pylint==2.9.3
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
wrapt==1.12.1
